uart_tester/requirements.txt
astroid==2.4.2
audio2numpy==0.1.2
certifi==2020.6.20
cffi==1.14.3
colorama==0.4.4
cycler==0.10.0
ffmpeg==1.4
isort==5.6.4
keyboard==0.13.5
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
matplotlib==3.3.2
mccabe==0.6.1
numpy==1.19.4
Pillow==8.0.1
pycparser==2.20
pylint==2.6.0
pyparsing==2.4.7
pyserial==3.4
python-dateutil==2.8.1
scipy==1.5.3
six==1.15.0
sounddevice==0.4.1
toml==0.10.2
wrapt==1.12.1
